+++ lib/info_kit.dart
@@ -7,3 +7,8 @@ export 'src/domain/mode.dart';
 export 'src/domain/os.dart';
 export 'src/domain/platform.dart';
 export 'src/domain/size.dart';
+
+export 'src/domain/platform_data.dart';
+export 'src/domain/size_data.dart';
+
+export 'src/data/flavor.dart';