+++ lib/src/locale/locale_default.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+class LocalePlatform {
+  static String? get locale => Platform.localeName.substring(0, 2);
+}