+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.1.0+2
+
+### Changed
+* Updated dependencies
+* Updated documentation
+* Get device locale on web
+* Device sizes are set and not dynamic
+
+### Added
+* `PlatformData` class and `SizeData` class
+* Get device screen size
+* Method to get store URL for app
+
 ## 1.0.0+1
 
 ### Changed